+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - you can add your homebrew features
 - live manager All Sources fixed
 - easier to ready things on live manager
-- You can preview damage type and change it
+- live manager You can preview damage type and change it
+- live manager You can change the critical now
 
 # 0.1.6
 - compendium-stats shows experience